requirements.txt
beautifulsoup4==4.5.1
billiard==3.5.0.2
boto==2.43.0
celery==4.1.0
django-annoying==0.10.3
django-autocomplete-light==3.3.2
django-colorfield==0.1.15
django-constance[database]==2.4.0
django-cors-headers==2.1.0
django-crispy-forms==1.7.2
django-extensions==2.1.6
django-notifications-hq==1.5.0
django-oauth-toolkit==1.0.0
django-phonenumber-field==1.3.0
django-recurrence==1.8.2
django-registration==3.0
django-reversion-compare==0.8.4
django-reversion==2.0.10
django-storages==1.5.1
django-timezone-field==3.0
django-tinymce==2.8.0
django==2.1.7
django_celery_beat==1.1.0
django_celery_results==1.0.1
django_filter==1.1.0
factory_boy==2.8.1
furl==1.2.1                         
google-api-python-client==1.6.4
graphene-django==1.3
graphene==1.4.1
graphql-core==1.1
graphql-relay==0.4.5
gunicorn==19.7.1
html2text==2016.9.19
kombu==4.1.0
mandrill==1.0.57
model-mommy==1.6.0
Pillow==3.4.2
psycopg2-binary==2.7.4
pysftp==0.2.9
python-dateutil==2.6.0
python-logstash==0.4.6
python3-keyczar==0.71rc0
pytidylib==0.3.2
pytz==2017.2
raven==6.9.0                        
requests==2.13.0
retry==0.9.2
simple-salesforce==0.73.0
stripe==1.59.0
tblib==1.3.2                        
twilio==6.28.0
unittest-xml-reporting==2.1.0
zeep==2.5.0
holidays==0.9.10
nameparser==1.0.4
exponent-server-sdk==0.2.0
django-import-export==1.2.0
